fix: 修复 v0.4.3 部署工作流中的 Dart 编译错误 (#26)
修复以下导致 Android / Web 构建失败的编译问题：

1. runtime_scope.dart / page_lifecycle.dart / event_input_injector.dart /
   binding_resolver.dart：修正相对导入路径（../../data/models → ../../../data/models）。
2. database_executor_io.dart：sqflite 与项目 DatabaseExecutor 命名冲突，
   通过 `hide DatabaseExecutor` 隐藏 sqflite 的同名类。
3. function_def.dart：补充 `variable_ref.dart` 导入，以使用 VariableRef / VariableSource。
4. node_executors.dart：
   - 移除不存在的 `JsEvalResult.isUndefined`，改用 `rawResult`；
   - 在 date_op 节点中补充解析 `params['b']`，修复 `b` 未定义。
5. function_editor_screen.dart：修复 Node? 空安全提升与 DragUpdateDetails 参数类型不匹配。
6. node_widget.dart：borderWidth 转为 double，满足 Border.all(width:) 类型要求。

+++ lib/features/node_graph/node_widget.dart
@@ -182,7 +182,7 @@ class _NodeCardState extends State<NodeCard>
                 borderRadius: BorderRadius.circular(12),
                 border: Border.all(
                   color: borderColor,
-                  width: borderWidth,
+                  width: borderWidth.toDouble(),
                 ),
                 boxShadow: [
                   BoxShadow(
